+++ src/main/java/cz/vsb/minibank/infrastructure/json/dto/JsonTransfer.java
@@ -1,5 +1,7 @@
 package cz.vsb.minibank.infrastructure.json.dto;
 
+import java.math.BigDecimal;
+
 /**
  * JSON representation of a transfer including authorization metadata.
  */
@@ -8,16 +10,20 @@ public class JsonTransfer {
     public int sourceAccountId;
     public Integer beneficiaryId;
     public String targetIbanSnapshot;
-    public double amount;
+
+    /** Held as the domain holds it. See {@link JsonAccount#balance} for why it is not a double. */
+    public BigDecimal amount;
+
     public String currency;
 
     /**
      * The fee actually charged, or null while the transfer has not settled.
      *
-     * Boxed, so "charged nothing" and "not charged yet" stay different facts - a ZeroFeePolicy
-     * really does charge 0.00 and that has to survive the round trip as 0.00, not as absent.
+     * Nullable on purpose, so "charged nothing" and "not charged yet" stay different facts - a
+     * ZeroFeePolicy really does charge 0.00 and that has to survive the round trip as 0.00 rather
+     * than as absent.
      */
-    public Double fee;
+    public BigDecimal fee;
 
     /** The customer's own reference, at most 140 characters. Null when none was given. */
     public String message;
